refactor: creating request and formatting response moved into new functions
