+++ src/request.rs
@@ -1,3 +1,6 @@
+use async_std::io::BufReader;
+use futures::{AsyncBufReadExt, AsyncRead, AsyncReadExt};
+
 use crate::handler::InternalResult;
 
 pub type Method = http::Method;
@@ -19,6 +22,9 @@ impl Request<()> {
         }
     }
 }
+
+const BAD_REQUEST: &str = "400 Bad Request";
+const HTTP_VERSION_NOT_SUPPORTED: &str = "505 HTTP Version Not Supported";
 
 impl<T> Request<T> {
     pub fn new(value: T) -> Self {
@@ -60,6 +66,106 @@ impl<T> Request<T> {
 
     pub fn headers(&self) -> &HttpHeaderMap<HttpHeaderValue> {
         self.request.headers()
+    }
+}
+
+impl Request<()> {
+    #[inline]
+    pub async fn from_stream(
+        mut stream: &mut (impl AsyncRead + Unpin),
+    ) -> std::result::Result<Request<String>, Box<dyn std::error::Error + Send + Sync>> {
+        let mut buf_reader = BufReader::new(&mut stream);
+        let mut first = String::with_capacity(1024);
+        buf_reader
+            .read_line(&mut first)
+            .await
+            .map_err(|_| BAD_REQUEST)?;
+
+        let mut request_builder = Request::builder();
+
+        first.pop();
+        first.pop();
+        let fl = first;
+
+        let mut splitted_fl = fl.split(' ');
+        let method = match splitted_fl.next() {
+            Some(mtd) => Method::try_from(mtd).map_err(|_| BAD_REQUEST)?,
+            None => Err(BAD_REQUEST)?,
+        };
+
+        let method = match method {
+            Method::GET => method,
+            Method::PUT => method,
+            Method::POST => method,
+            Method::DELETE => method,
+            Method::OPTIONS => method,
+            Method::HEAD => method,
+            Method::TRACE => method,
+            Method::PATCH => method,
+            Method::CONNECT => method,
+            _ => Err(BAD_REQUEST)?,
+        };
+
+        let uri = match splitted_fl.next() {
+            Some(mtd) => Uri::try_from(mtd).map_err(|_| BAD_REQUEST)?,
+            None => Err(BAD_REQUEST)?,
+        };
+
+        match splitted_fl.next() {
+            Some("HTTP/1.1") => (),
+            _ => Err(HTTP_VERSION_NOT_SUPPORTED)?,
+        };
+
+        let mut content_length = 0usize;
+
+        loop {
+            let mut line = String::with_capacity(100);
+            buf_reader
+                .read_line(&mut line)
+                .await
+                .map_err(|_| BAD_REQUEST)?;
+
+            line.pop();
+            line.pop();
+
+            if line.is_empty() {
+                break;
+            }
+
+            let splitted_header = line.split_once(':');
+            match splitted_header {
+                Some((header, value)) if http::header::CONTENT_LENGTH == header => {
+                    request_builder = request_builder.header("Content-Length", value);
+                    content_length = value
+                        .trim()
+                        .parse::<usize>()
+                        .unwrap();
+                }
+                Some((header, value)) => {
+                    match (
+                        HttpHeaderName::try_from(header.trim()),
+                        HttpHeaderValue::try_from(value.trim()),
+                    ) {
+                        (Ok(header), Ok(value)) => {
+                            request_builder = request_builder.header(header, value);
+                        }
+                        _ => Err("400 Bad Request")?,
+                    }
+                }
+                None => Err("400 Bad Request")?,
+            }
+        }
+
+        let mut body_string = vec![0u8; content_length];
+        buf_reader
+            .read_exact(&mut body_string)
+            .await
+            .map_err(|_| BAD_REQUEST)?;
+
+        Ok(request_builder
+            .uri(uri)
+            .method(method)
+            .body(String::from_utf8(body_string)?))
     }
 }
